inferできる

+++ catkin_ws/src/image_test/src/infer.cu
@@ -6,17 +6,14 @@
 #include <opencv2/opencv.hpp>
 #include <chrono>
 #include <ros/ros.h>
-#include "utils.h"
+#include <utils.h>
 
-using namespace std;
 using namespace nvinfer1;
 
-class Logger : public ILogger
-{
-  void log(Severity severity, const char * msg) override
-  {
+class Logger : public ILogger {
+  void log(Severity severity, const char * msg) override {
     if (severity != Severity::kINFO)
-      cout << msg << endl;
+      ROS_INFO("%s", msg);
   }
 } gLogger;
 
@@ -25,19 +22,25 @@ ICudaEngine *engine;
 IExecutionContext *context;
 int inputBindingIndex, outputBindingIndex;
 int inputHeight, inputWidth;
+Dims inputDims, outputDims;
 bool is_initialized = false;
 void *bindings[2];
 
-void setup(string planFilename, string inputName, string outputName) {
+// pointers
+size_t numInput, numOutput;
+float *inputDataHost, *outputDataHost;
+float *inputDataDevice, *outputDataDevice;
+
+void setup(std::string planFilename, std::string inputName, std::string outputName) {
   ROS_INFO("setup");
-  ifstream planFile(planFilename);
+  std::ifstream planFile(planFilename);
   if(!planFile.is_open()) {
     ROS_INFO("cannot get plan file");
     is_initialized = false;
   } else {
-    stringstream planBuffer;
+    std::stringstream planBuffer;
     planBuffer << planFile.rdbuf();
-    string plan = planBuffer.str();
+    std::string plan = planBuffer.str();
 
     runtime = createInferRuntime(gLogger);
     engine  = runtime->deserializeCudaEngine((void*)plan.data(), plan.size(), nullptr);
@@ -46,45 +49,58 @@ void setup(string planFilename, string inputName, string outputName) {
 
     inputBindingIndex = engine->getBindingIndex(inputName.c_str());
     outputBindingIndex = engine->getBindingIndex(outputName.c_str());
-    Dims inputDims, outputDims;
     inputDims = engine->getBindingDimensions(inputBindingIndex);
     outputDims = engine->getBindingDimensions(outputBindingIndex);
     inputHeight = inputDims.d[1];
     inputWidth = inputDims.d[2];
 
-    size_t numInput, numOutput;
     numInput = numTensorElements(inputDims);
     numOutput = numTensorElements(outputDims);
 
     // host
-    float *inputDataHost, *outputDataHost;
     inputDataHost = (float*) malloc(numInput * sizeof(float));
     outputDataHost = (float*) malloc(numOutput * sizeof(float));
     // device
-    float *inputDataDevice, *outputDataDevice;
     cudaMalloc(&inputDataDevice, numInput * sizeof(float));
     cudaMalloc(&outputDataDevice, numOutput * sizeof(float));
 
     is_initialized = true;
-    ROS_INFO("initialize finished");
+    ROS_INFO("initialize finished %d, %d", numInput, numOutput);
   }
 }
 
 void destroy(void) {
-  runtime->destroy();
-  engine->destroy();
-  context->destroy();
-
+  if(is_initialized) {
+    runtime->destroy();
+    engine->destroy();
+    context->destroy();
+    free(inputDataHost);
+    free(outputDataHost);
+    cudaFree(inputDataDevice);
+    cudaFree(outputDataDevice);
+  }
   is_initialized = false;
 }
 
 void infer(cv::Mat image) {
+  // cvの画像からcnnを走らせる
   ROS_INFO("get");
-  /* cvImageToTensor(image, inputDataHost, inputDims); */
+  cvImageToTensor(image, inputDataHost, inputDims);
+  preprocessVgg(inputDataHost, inputDims);
+  bindings[inputBindingIndex] = (void*)inputDataDevice;
+  bindings[outputBindingIndex] = (void*)outputDataDevice;
+
+  cudaMemcpy(inputDataDevice, inputDataHost, numInput * sizeof(float), cudaMemcpyHostToDevice);
+  context->execute(1, bindings);
+  cudaMemcpy(outputDataHost, outputDataDevice, numOutput * sizeof(float), cudaMemcpyDeviceToHost);
+  // output
+  ROS_INFO("%f %f %f %f", outputDataHost[0], outputDataHost[1], outputDataHost[2], outputDataHost[3]);
 }
 
 void test(void) {
   ROS_INFO("inside cu");
   cudaDeviceSynchronize();
 }
+
+
